+++ sdk/datafactory/Microsoft.Azure.Management.DataFactory/src/CHANGELOG.md
@@ -7,6 +7,8 @@
 - Added maxRowPerFile/fileNamePrefix to tabular format settings
 - Added support for AzureDatabricks delta lake
 - Update Rest Sink properties
+- Added support lzo in orc format
+- Added MongoDbAtlas Source Connector in DataFactory
 
 ## Version 4.11.0
 ###  Feature Additions